# server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bora-Boracle",
    "version": "0.0.1",
    "description": "Mini projet visant à valider nos compétences en BDD"
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "tags": [
    {
      "name": "Cocktail",
      "description": "API routes for cocktails"
    },
    {
      "name": "Orders",
      "description": "API routes for orders"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "parameters": [
          {
            "name": "Order",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "tags": [
          "Orders"
        ],
        "summary": "Create a new order",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/order/{Id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "type": "number",
            "required": true
          }
        ],
        "summary": "Get order by ID",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "type": "number",
            "required": true
          },
          {
            "name": "Order",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "summary": "Update order",
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Cocktail": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "number",
          "uniqueItems": true
        },
        "Name": {
          "type": "string"
        },
        "Price": {
          "type": "number"
        },
        "Photo": {
          "type": "string"
        },
        "Description": {
          "type": "string"
        }
      }
    },
    "CocktailOrderList": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "number",
          "uniqueItems": true
        },
        "Order": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Order"
          }
        },
        "Cocktail": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Cocktail"
          }
        },
        "Delivered": {
          "type": "boolean"
        },
        "Ready": {
          "type": "boolean"
        },
        "OrderedAt": {
          "type": "string"
        }
      }
    },
    "TableInfo": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "number",
          "uniqueItems": true
        },
        "Number": {
          "type": "number"
        },
        "Location": {
          "type": "string"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "number",
          "uniqueItems": true
        },
        "PeopleName": {
          "type": "string"
        },
        "Complete": {
          "type": "boolean"
        },
        "TableInfo": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/TableInfo"
          }
        }
      }
    },
    "Orders": {
      "type": "object",
      "properties": {
        "orders": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Order"
          }
        }
      }
    }
  }
}